requirements.txt
Django==4.2.9
Django==4.2.16
djangorestframework==3.15.2
